feat(game): move bottom ui buttons to board window header
