+++ src/constants/size.ts
@@ -8,3 +8,5 @@ export const MIN_MINES = 10;
 
 export const GAME_AREA_PERCENTAGE = 88;
 export const GAME_BOARD_SIZE = `${GAME_AREA_PERCENTAGE}vmin`;
+
+export const GAME_BOARD_HEADER_SIZE = '4vmin';
